Extracted SkillTag and ProjectMemberCard from ProjectManagementDialog
