+++ frontend/src/utils/formatTime.ts
@@ -20,3 +20,11 @@ const formatTime = (timestamp: string) => {
 };
 
 export default formatTime;
+
+export const formatDate = (dateString: string) => {
+    return new Date(dateString).toLocaleDateString('en-US', {
+        year: 'numeric',
+        month: 'short',
+        day: 'numeric'
+    });
+};
